AP-15867: Fix UI behaviour on restore to default
Previously, when restoring to defaults in the preference page, the logic
would update the state of the input fields based one the "use defaults"
value in the current configuration. This is problematic if "use
defaults" is unchecked, that setting is saved via "apply" and then the
user clicks "restore defaults". This commit makes the following change:
When the "Restore Defaults" button is pressed, the UI is updated with a
state to always use defaults (not access the current settings).

AP-15867 (Columnar Table Preferences UI Glitch)