AP-15312: Save metadata in the schema
This will be needed for writing compatible files with Python. With the
pyarrow API it is not possible to set the metadata in the footer but it
is easy to set the metadata in the schema.

AP-15312 (Implement Memory-Mapped File based reading of Arrow files)


(This commit was moved from knime-core-arrow. Original id: 2d652839a9ae3fb51797f023d66963d927d01d98)