AP-17059: Not assume padding of 4 when setting struct validity buffer
When the vector is only allocated for a few elements the validity buffer
can be smaller than 4. If we always use ints to set the validity buffer
we access memory outside of the buffer for the last bytes.
This is fixed by checking if more than 4 bytes are still to be set and
if not, the validity is set by using single bytes.

Also, note the change in the tests which now allows for buffer
allocations that are not a multiple of 4. Without the fix, this fails in
multiple tests.

AP-17059 (Boolean Sets with only a few rows cause unsafe memory access in fast tables)


(This commit was moved from knime-core-arrow. Original id: 87335c48d1bc7bf8a39e070fac77917d921c67bc)