AP-17192: add test for calling expand during asynchronous serialization
In CachedWriteData we are performing most operations on the write
delegate asynchronously in CompletableFutures, but expand was not
asynchronous up to now. It is a special case because we do not just need
to tell the delegate to expand its data buffer, but we also need to
increase the size of the heap cached data. The new test ensures
that this works correctly.

AP-17192 (Columnar backend memory issues)