AP-17192: flush ColumnarRowContainer if required to also work in loops
In loops, the node context has a forceSynchronousIO flag set because
the loop end node expects each row to be fully written before it calls
reset and starts the next loop iteration. We need to make sure we flush
the caches after each written row to fulfill this contract. In all other
cases we only flush when finishing writing the RowContainer.

AP-17192 (Columnar backend memory issues)