AP-17001: delete file after closing batch store, moved responsibility to user
We streamlined the side effects of ArrowBatchStore.close() to behave
more like a file handle: it opens a file for writing, but does no longer
delete the file when closing the BatchStore. Thus we need to delete
the file whenever the BatchStore is closed and we don't need the file
any more.

AP-17001 (Refactor Batches and improve naming of length/size/capacity)