AP-17192: Run ObjectCache and CachedWriteData tasks on different executors, adjust tests
The ObjectCache performs asynchronous insertion of data into the
underlying cache, while CachedWriteData asynchronously serializes values and
passes them on to the underlying delegate. These two operations could
potentially be blocking each other if running on the same thread,
so we split them into sparate thread pools (as it was working a while ago).

The tests were controlling the thread to check the behavior of
ObjectCache and CachedWriteData, and now they need to control both
executors.

AP-17192 (Columnar backend memory issues)