AP-17192: call ArrowVector expand on calling thread and not in future
Apparently in the Arrow back-end calling expand on two vectors
concurrently leads to undefined behavior, at least doing so did
cause corrupted data in the topNodesColumnar workflow.
Because we do not want the back-end to deal with these kind of
concurrency issues introduced by the asynchronous heap-cache
serialization, we block during expand() and invoke expansion of the
delegate in the caller thread.

AP-17192 (Columnar backend memory issues)