AP-17327: Implement dictionary encoding using StructData
Because the Arrow implementation of dictionaries is missing the handling
of replacement or delta dictionaries in C++ and Python, we decided to
implement dictionary encoding on our own: we use a struct of (dictKey,
hash, Optional<Value>) at each position in the array. The optional value
will only be filled once for each key per batch, meaning that all values
that occur in a batch are also stored in the batch.


(This commit was moved from knime-core-arrow. Original id: 5c4d260cd19f76d867128ea0ce4ed4a473804832)