AP-17327: Leave dictionary encoding implementation to the back-end
We provide an interface that adds general methods to set or retrieve
the dictionary key used at each position in the data, but otherwise
leave the implementation to the back-end. (Arrow now uses structs).

To prepare for table-wide dictionary entry caching, we wrap all
dictionary encoded data coming from the back-end and inject a
DictElementCache, which does not do anything for now.

Added a test implementation for DictEncodedData by wrapping structs.