AP-16402: fix deadlock in ReadDataCache by using a smaller scoped synchronize block
In ReadDataCache.close() we invoke flush(), which waits for all asynchronous
writing operations to finish. However, close() was synchronized, and in
the writing operations there was also a block that synchronized on
ReadDataCache. Thus, if close was called while not all writing operations
were finished, this setup was always running into a deadlock.

The synchronization block inside the write operations was only necessary
to check whether the cache was closed already. This is replaced by and
AtomicBoolean flag now.

Added a test to check for deadlock on ReadDataCache.close while writing,
which would hang before this change.