AP-16402: make ObjectCache.close() synchronous again to behave more like any file handle or stream
Letting the ObjectCache continue to write asynchronously after calling
close made it impossible to know when the underlying batch store
delegate was finally closed. But we need this information because
we want to read / delete the file written by the BatchStore back end.

Remove test constructs necessary to deal with asynchronous ObjectCache.close