AP-16402: Do not assert equal batch length for the last batch, can be smaller
In our tests we explicitly construct data to be 1 element longer than the max
batch length. When writing this data to the TestBatchStore we checked
whether all batches have the same length, which is obviously false here.

Due to changes in the concurrency models of ObjectCache and ReadDataCache
we started running into this error quite frequently. Probably we were
never waiting for writing to finish in the tests and thus the error did
not pop up before, or it is showing up due to more sensible error
propagation. Anyways, the assumption was wrong and is fixed now.