AP-16402: Use Phaser to wait for all asynchronous serialization before closing ObjectCache
In ObjectCache.close() we already wait for all asynchronous operations
of the ObjectCacheWriter, however that one dispatches asynchronous
serialization in CachedWriteData which might still run even after the
writer tasks are done. We now create a Phaser in the ObjectCache and let
the CachedWriteData async tasks register themselves there, and they
deregister on completion. This way the ObjectCache can wait for all
async serialization tasks to complete (with early-termination in case
of cancellation) and make sure it's safe to close the delegate.