AP-16402: Fix crash by making sure we do not invoke Arrow's reAlloc() concurrently
reAlloc() of the underlying Arrow vector can also be called during
serialization of variable length data, e.g. in
ArrowBuf.ArrowBufDataOutput#ensureCapacity. If this happens at the same
time as we call expand() on this vector, we can create a corrupt state
of the ArrowVector, which apparently caused frequent crashes on Windows.