AP-17320: Swap dict encoding and caching layers
If we decide to enable both, caching and dictionary encoding for String and
VarBinary data, then it makes more sense to cache the underlying data
(the dictionary value column) because that is already the compressed
data. Thus, we move the dictionary encoding layer above the cache
layers, further away from the back end closer to the user.

To achieve this, we change the DictEncodedBatchStore to a
BatchWritableReadable so it can use the CachedBatchWritableReadable as
delegate.