AP-17500: Make dictionary key type configurable
Which adds a lot of generics to the code, but allows us to use smaller
range keys if we know that there is only a limited number of values.

There are now also three key generators for byte, int and long keys
which check that they do not generate a key collision.

Added test coverage for all three key types.

AP-17500 (Make key type of struct dict encoding configurable)