AP-17624: Fix ArrowSchema conversion to ColumnarSchema for StructDictEncoded fields
When coming from Python we noticed a difference between the ArrowSchema
and the ColumnarSchema. The reason was that DictEncoded Strings did not
have a StringDataSpec but a StructDataSpec with DictEncoding traits.

AP-17624 (Bump ArrowColumnarDataFactory versions)


(This commit was moved from knime-core-arrow. Original id: 2ee5863c17f70ce6a16574cea82d0fb598d32b24)