AP-17790: Throw exception if mapped batch is removed twice
A mapped batch can only be removed once. If it is removed twice it is
an implementation issue of the code using the batch. To make it easier
to find such implementation issues we now throw an exception if this
happens.

AP-17790 (NPE when releasing MappedByteBuffer)


(This commit was moved from knime-core-arrow. Original id: 5d18891c76a9a25d3e9a87901207db14acf33198)