AP-17914: Flush the caches in correct order
The heap cache writes to the small table cache which writes to the
column data cache. Therefore they also need to be flushed in this order,
otherwise the column data cache might asynchronously write to disk which
is a problem if we want to save the table which requires that all data
has been written by the time flush returns.

AP-17941 (Don't flush on finish)