AP-18296: Use CursorTracker to track RowCursor in ColumnarRowReadTable
Previously this functionality was baked into the DefaultRowCursor but as
we now need the same functionality for (Lookahead)Cursor, we extracted
the logic to reduce code duplication. The alternative would be to bake
the resource leak detection into Cursor but this would require
org.knime.core.columnar to know about ResourceLeakDetector.

AP-18296 (Memory leaks in loops with wide tables)