AP-18792: Do not rely on free memory size on MacOS because it is not reliable
The RAM usage strategy on macOS is different from other OSes in that macOS uses
the RAM as cache and only drops things from that cache if needed. Thus, if a Mac
has been running for some time, the amount of "free" memory will be close to zero.
Instead, we now use half the size of the RAM minus JVM memory as "free memory" heuristic.

AP-18792 (Cache size heuristics do not make sense on MacOS)