AP-18871: Simplify concurrency logic
The m_pendingBatchIdx is used in BufferedDataContainerDelegate to ensure
that the batches are written by the async ContainerRunnables in the correct
order. Basically, each ContainerRunnable does the spec validation and
RowKey checking for its designated batch, unconditionally. The actual
writing (to the Buffer) of batches that complete these checks out of order
is then only done by one of the ContainerRunnables, taking all checked
batches so far (also from other ContainerRunnables) in order.

The ContainerRunnables of ColumnarDataContainerDelegate don't do any
checks. They just write to the underlying storage. Therefore it seems not
useful to have the same logic, because there is no work that could be
completed out of order by concurrent ContainerRunnables. The whole
m_pendingBatchIdx logic is overkill. This is simplified here to having
one async task after another, chained by CompletableFuture.thenRunAsync().