AP-19695: Check the singleton store factory for canLoad
If the singleton column store factory cannot be initialized, loading the
table will fail because the constructor of the
AbstractColumnarContainerTable will fail. Now, we notify the caller
earlier that we won't be able to load any table.

AP-19695 (WF created with columnar table backend cannot be saved in KNIME without columnar table backend)