AP-20233: Introduce test mode to rewrite tables in workflow tests
Our workflow tests are often stored with the row-based backend in a
partially executed state. Therefore the input tables to the
ColumnarTableBackend are often non-columnar in that scenario. Since our
goal is to test the behavior in the ideal scenario of every table being
columnar, we rewrite the tables if they aren't but only when running
workflow tests as indicated by a corresponding system property.