Do not wait for writing in the ReadDataCache evictor
The data is retained anyway. Therefore, we do not wait for writers to
finishe when data is released from the cache. However, we cannot trust
the cache anymore to include the data that we are currently writing
while reading.

This fixes a deadlock that could happen if all writers wait for a memory
alert that itself waits for a writer.