AP-21661: Adapt knime-core-table API changes
Make ColumnarVirtualTable methods public.

Fix TableTransformNodeSettingsPersistor for MapTransformSpec:
Persistence now works for both ColumnarMapperWithRowIndexFactory and
ColumnarMapperFactory.

For deserialization, ColumnarMapperWithRowIndexFactory is wrapped as a
ColumnarMapperFactory, using WrappedColumnarMapperWithRowIndexFactory.
For serializaton, if the ColumnarMapperFactory is a
WrappedColumnarMapperWithRowIndexFactory, the wrapped
ColumnarMapperWithRowIndexFactory is serialized instead of the wrapper.

TableCasterFactory is now a ColumnarMapperFactory, because luckily it
doesn't use the RowIndex. TableCasterFactories with the previous version
of TableTransformNodeSettingsPersistor will thus be correctly deserialized,
with out needing to insert ROWINDEX TableTransforms.