AP-23793: revise ColumnarRearranger
Rewrite createAppendTable() to use the appendMap() virtual table method
instead of map() and append(). This prevents unnecessarily instantiating
copies of the source table when a lot columns are modified.

A new virtual table convenience method replaceMap() has also been added
which allows to modify a column in place (via appendMap(), a column
permutation, and dropping the original column).